Logic of CachingFilter changed.
Original: VTK filters that should support caching were supposed to be derived from the standard version of the filter and from CachingFilter plus they were supposed to implement methods staticRequestData and RequestData

New: VTK filters that should support caching are supposed to be derived only from CachingFilter passing the standard version of the filter as type parameter for CachingFilter. Methods staticRequestData and RequestData are no longer supposed to be implemented.